[refactor] TestController renamed to ExampleDBController
